chatbot_rewe_group/services/az_functions/llm/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core==1.34.0
azure-data-tables==12.7.0
azure-functions==1.23.0
azure-identity==1.23.0
azure-keyvault-secrets==4.9.0
azure-storage-blob==12.25.1
graphiti-core==0.17.6
langchain==0.3.25
langchain-openai==0.3.19
langchain-core==0.3.63
langgraph==0.4.8
openai==1.93.0
pydantic==2.11.7